+++ tests/testthat/test_latest_flights_date.R
@@ -11,7 +11,7 @@ test_that("latest_flights_date", {
 
   dates1 <- latest_flights_date()
 
-  testthat::expect_true(class(latest_flights_date()) == "numeric")
+  testthat::expect_true(class(latest_flights_date()) == "integer")
   testthat::expect_true(length(dates1) > 0)
   testthat::expect_error(latest_flights_date(999))
 })
